modified the (PUT and DELETE) updelusermanagement method under the  user_management_view.py

+++ DOEAssessmentApp/DOEAssessmentApp/DOE_views/user_management_view.py
@@ -180,18 +180,50 @@ def updateAndDelete():
                         result = [{col: getattr(d, col) for col in colsusermanagement} for d in data]
                         return make_response(jsonify({"data": result[0]})), 200
                     elif request.method == 'PUT':
-                        user_name = res['empname']
-                        user_emppassword = generate_password_hash(res['EmployeePassword'])
-                        data.first().emppasswordhash = user_emppassword
-                        data.first().empname = user_name
-                        db.session.add(data.first())
-                        db.session.commit()
-                        return make_response(jsonify({"msg": f"UserManagement {user_name} successfully updated."})), 200
+                        res = request.get_json(force=True)
+                        user_role = res['emprole']
+                        if data.first().emprole == 'admin':
+                            count_user_with_admin_role = Companyuserdetails.query.filter_by(emprole=user_role).count()
+                            if count_user_with_admin_role > 1:
+                                data.first().emprole = user_role
+                                db.session.add(data.first())
+                                db.session.commit()
+                                return make_response(
+                                    jsonify(
+                                        {"msg": f"UserManagement successfully updated with role {user_role}."})), 200
+                            else:
+                                return make_response(
+                                    jsonify({
+                                        "msg": "Please assign a user to admin role before changing the current admin's role"})), 200
+                        else:
+                            data.first().emprole = user_role
+                            db.session.add(data.first())
+                            db.session.commit()
+                            return make_response(
+                                jsonify(
+                                    {"msg": f"UserManagement successfully updated with role {user_role}."})), 200
+
                     elif request.method == 'DELETE':
-                        db.session.delete(data.first())
-                        db.session.commit()
-                        return make_response(jsonify({"msg": f"UserManagement with ID {row_id} "
-                                                             f"successfully deleted."})), 204
+                        res = request.get_json(force=True)
+                        user_role = res['emprole']
+                        if data.first().emprole == 'admin':
+                            count_user_with_admin_role = Companyuserdetails.query.filter_by(emprole=user_role).count()
+                            if count_user_with_admin_role > 1:
+                                data.first().emprole = user_role
+                                db.session.delete(data.first())
+                                db.session.commit()
+                                return make_response(
+                                    jsonify(
+                                        {"msg": f"UserManagement with ID {row_id} "
+                                                f"successfully deleted."})), 204
+                            else:
+                                return make_response(jsonify({
+                                    "msg": "Please assign a user to admin role before deleting the current admin's role."})), 204
+                        else:
+                            db.session.delete(data.first())
+                            db.session.commit()
+                            return make_response(jsonify({"msg": f"UserManagement with ID {row_id} "
+                                                                 f"successfully deleted."})), 204
             else:
                 return make_response(jsonify({"msg": resp})), 401
         else:
